(build) Update dependencies and fix tests
Weird gotcha is that Jooby seems to have changed the generated service suffix to add Htmx for some reason.

+++ code/functions/language-processing/java/nu/marginalia/keyword/model/DocumentKeywordsBuilder.java
@@ -230,7 +230,7 @@ public class DocumentKeywordsBuilder {
         wordToMeta.forEach((word, meta) -> {
             sb.append(word)
                     .append("->")
-                    .append(WordFlags.decode((byte) meta.longValue()))
+                    .append(WordFlags.decode((byte) meta))
                     .append(',')
                     .append(wordToPos.getOrDefault(word, new IntArrayList()))
                     .append(' ');
